Fixed max-k-step and k-step
If one specifies `k-step` value that is equal or bigger than
`max-k-step` value (default or explicitly set), ESBMC only
tries `k=1` and then stops. This combination of values for
`k-step` and `max-k-step` should be caught by ESBMC and
reported at least as an error as it can not be the user
intention.